Allow KubeAdm config customizing and port forwarding.
Instead of hard coding the kubeadm.conf settings as part of the
dind-cluster.sh script, we'll use a template file and customize
it with the known settings needed based on the IP mode and CNI
mode.

The user can customize the template as desired, and rebuild the
DinD image (build/build-local.sh) and set DIND_IMAGE to use these
changes.

For minion nodes, allow port forwarding, using the EXTRA_PORTS
configuration setting.

As part of setting the static route to the NAT64 server for IPv6,
set the docker flag to remove the temp container used for that
command.

Found an issue in IPv6 mode, where kube-proxy has cluster CIDR set
to the IPv4 address and prefix of docker0 interface, which causes
issues when used by ip6tables. For IPv6 mode, the cluster CIDR
should be an empty string.

To handle this, the dind-cluster.sh script touches a file on the
kube-master node, so that the wrapkubeadm script can tell how to
set the cluster CIDR (based on IP mode).