Use separate pod networks, instead of nesting under dindnet.
There are several parts to this change. First, we'll use the existing
POD_NETWORK_CIDR variable to indicate the cluster for pod networks.
Each node will have a subnet in that cluster, that uses the node's ID
as the discriminator. For example, 10.244.0.0/16 (the default), will
define 10.244.1.0/24 for kube-master, 10.244.2.0/24 for the first
minion, etc.

Second, the POD_NETWORK_CIDR will be broken into a prefix and a size,
that is passed to dindnet, to properly provision the CNI config file
for bridge plugin. For IPv6, the code will pad the prefix with zeros,
so that it is fully qualified and then append the node's ID, giving a
proper prefix that can be used to create the GW IP and network address.

There are a few cases for the forming of the prefix in IPv6. The simple
case is where the POD_NETWORK_CIDR of "fd00:40::/72" is converted into a
prefix of "fd00:40:0:0:" and the node ID (e.g. 2) is added, to give a prefix
of "fd00:40:0:0:2". On the node, this would give a pod network value of
"fd00:40:0:0:2::/80".

Another case is when the user specifies a fully qualified POD_NETWORK_CIDR
of "fd00:10:20:30:4000::/72". This will be truncated to the prefix value of
"fd00:10:20:30:40", and then with the node ID added, we get pod network of
"fd00:10:20:30:4002::/80" on a node with ID 2.

The last case, is with a POD_NETWORK_CIDR of "fd00:40::/64", where the
prefix will be "fd00:40:0:0:", and the node ID will be placed in the upper
byte of the next word of the prefix. On a node with ID 2, this gives a pod
network of "fd00:40:0:0:0200::/72".

Third, the CNI config uses ipMasq (for outside access), isDefaultGateway
(to give dind0 interface an IP), and no longer specifies the minimum
and maximum host IP for the pod network, default route, or route to DNS64
network via the NAT64 (as that is added to the eth0 interface).

Fourth, for IPv4, we'll assume that the POD_NETWORK_CIDR has a /16 size
and pod networks are /24.  For IPv6, pod networks will be 8 bits smaller
than the cluster size as well.

Fifth, because the pod networks are no longer nested inside of the dindnet,
static routes are added on each node, so that pods can access pods on
other nodes.

Sixth, for E2E tests we need to have hairpin mode enabled in the CNI config
file. Created USE_HAIRPIN environment variable, which will default to false
for IPv4 (for backwards compatibility, as there is some indication of an
issue when using hairpin mode and Virtlet Networking), and will default to
true for IPv6.