Adding a default config.sh for stable branch
Recently a stable branch was added for CI testing.
The scripts in the fixed directory were updated also, however that
only works for local clusters, they are bypassed when using GCE.
This change adds a default config file for that branch that can
be used by the CI test script when using GCE.
The config file can be copied over, then gce-setup.sh can be called
and the defaults for that branch will be used.
It is preferred to have the defaults in the kubeadm-dind-cluster repo
than in the CI repo.