Force MAC on pod network bridge.
In running E2E tests, it was observed that during the testing, the MAC
on the dind0 bridge was changing. This was causing, in some cases, the
kube-dns pod to fail health checks (lose TCP communication on the bridge
for that pod).

Research shows that, if a MAC is not assigned to a virtual bridge, then
the bridge will get the lowest MAC of all enslaved interfaces (dynamically).
This implies that the MAC could also change, and in this case, if a test
pod has a lower MAC than kube-dns (which is the first pod added to the
dind0 bridge at startup), then the MAC of the bridge would change.

This does not appear to be a problem for IPv4, where the bridge has a MAC
address assigned by the CNI plugin.

With the latest CNI plugin, v0.7.1, the plugin will set the MAC for
bridges created. It will adopt existing bridges (like dind0), but will
not ensure they have a MAC.

I did try changing the sequencing of the dindnet script to first create
the CNI config, and then create the bridge, but it appears that CNI does
not get involved in the bridge creation at that time (it does later).

As a solution, this commit will create a dummy veth pair, add one of the
interfaces into the bridge, and then set the bridge's MAC to the MAC of
the interface (reading up on this indicates that we must use a MAC of an
enslaved interface).

Now, dind0 has a MAC address, and since it was explicitly set, there is no
dynamic override of the address. During E2E tests, kube-dns health check
responses make it back to the kubelet, and it no longer gets restarted.

NOTE: This commit uses CNI v0.7.1, which is found in a different repo
location. We need this newer version, as v0.6 does not have IPv6 support,
and caused other issues (like IP6tables not being set up correctly).

NOTE: Tested IPv4 and IPv6 on local and GCE platforms