IPv6: Allow AAAA record use
By default, DNS is set up to ignore (IPv6) AAAA records. As a result,
whenever a remote site is accessed, only the (IPv4) A record is used,
and a synthesized IPv6 address formed. This means that NAT64 will be
used for all external accesses, even when a site supports IPv6.

This is fine for environments where the cluster only has IPv4 access,
like GCE currently. However, if the cluster will have IPv6 access too,
the configuration can be modified to use AAAA records in DNS lookups,
and then remote sites with IPv6 addresses can be directly accessed
(without using NAT64).

This change provides a knob to allow enabling use of AAAA records. In
addition to modifying the DNS configuration, ip6tables rules are
created to perform SNAT for packets coming from the container nodes
and to forward external packets to the kubeadm-dind-net to get to the
pods.

The boolean environment variable, DIND_ALLOW_AAAA_USE, is defined for
this capability. If set to any value, it will be enabled (so false or
zero are considered "set", meaning true). If unset (the default for
backwards compatibility) AAAA records will not be used.

For IPv4, if this is set, a warning message will be displayed and it
will be ignored (it only applies to IPv6). For IPv6, if running on
GCE and this is set, the run will be aborted, as GCE doesn't support
native IPv6 outside of the cluster.

Note: This commit also modifies GCE_SETUP to be a boolean, instead of
a string representation of a boolean, for consistency.

Fixes issue #147