Change default label and make it configurable
The default label has been changed to be different to the label on the
container image. This allows for easier container selection with keeping
the label on the image.

We als introduced a test, currently failing, which should drive out the
needed changes to make multiple instances of kubeadm-dind-cluster in
parallel working.