Explicitly set the context when calling kubectl
Previously the new context that was created for the dind cluster was
also set as default, and therefore all kubectl calls used to have it
available.

When supporting multiple dind clusters in parallel, the script still
creates a new context per cluster (same behaviour as before). However
now the script does not set the new context as default, which is why
it's required to set the context explicitly on kubectl calls.