Fix `dump`
Inside the cluster the configuration regarding contexts is different.

To connect to the APIServer from outside of the cluster you need to
specify the distinct context for the cluster in question. When
connecting to the APIServer from the inside one can just use the
defaults connection settings (no context, no user, ...).