Internal port for APIServer is not meant to be configured
We were under the assumption that the internal port the APIServer is
listening on is configurable. This turned out not to be true. Therefore
we rolled back a few changes regarding configuration options for the
internal port.

However, as part of this commit we introduced a read-only variable for the internal
port and changed the templating of kubeadm to use this variable. This
gives us the option to, in future, make the internal port configurable.
One thing we'd need to introduce for that is setting up the context
internally to use the non-default port.