Keep master & node name for default cluster
For the default cluster (default `DIND_LABEL`, `DIND_LABEL` not set),
the names for the master and the nodes will be `kube-master`,
`kube-node-<x>`.
For other clusters (non-default `DIND_LABEL`) the names of the master
and the nodes will be `kube-master-<hash>`, `kube-node-<hash>`, where
`<hash>` is the sha1 sum of the `DIND_LABEL`.