Fix restore bug
Until now, the portforwarding to the APIServer on restore assumed that the
APIServer is listening on the same port internally as the port
forwarding was setup for on the host (by default 8080).
So if a cluster was restored with `APISERVER_PORT=8081` the
portforwarding would forward the localhost:8081 to <APIServer>:8081,
however nothing was listening on <APIServer>:8081.

This change sets up the portforwarding to <APIServer>:8080, regardless of the
`APISERVER_PORT`. The APIServer will always listen on 8080, the
`APISERVER_PORT` config will only have infulence on the portforwarding.