Search for free port on the docker host
... not on the host, where the docker client is running.

Documentation suggests, it is left to the user to forward the correct
port from the docker client host to the remote docker host.
Let's keep that for now as is.

Also it is important to understand that this way to check for free ports
is super racy. We might want to have docker bind to a random port and
have a way to query that random port or an option to setup
portforwarding.