Missing use-context for non-multicluster
When not in multi-cluster mode, the set-context command must be
accompanied by the use-context command. Otherwise the host cannot
contact the cluster via API.

This change also renames the variable used to indicate whether the
docker daemon is running locally. Conditional tests can then
combine testing for GCE and remote docker daemon, as the both have
DOCKER_HOST set.

There are related/miscellaneous fixed as well...

In local docker mode, when trying to access the API from the host, even
with the context set, the connection was refused. The host indicates
that the API is at the master's IP on a specific port, however the host
netstat shows that it is listening on 127.0.0.1 on the port. Modified to
always use localhost or ::1, for IPv4 and IPv6, respectively.

For NAT64 we set a static route to for the IPv4 pool of the NAT64 server.
The remote docker case was not doing this route, so added in. Also, since
DOCKER_HOST is set for GCE or remote docker cases, the docker run command
can be used in all cases.

The DNS64 container used volume mounting, and created a config file on the
host in the /tmp area. This has several problems. First, a securty issue
with the temp file. Second, the file is copied over to a GCE container, but
not for a remote docker setup. To resolve this, the config is stored into
an environment variable and passed into the container at startup.

When running E2E tests, the API port is fixed to 8080. However, when running
under GCE, the API port is 8899 (and mapped to 8080 internally). This was
changed to use APISERVER_PORT for the test invocations.

Fixes issue: #181