IPv6: Addressing breakage and related issues
There are several issues found:

1) Every time an address is needed on the DinD network, a lookup
   is done to determine the CIDR. This is inefficient, and it
   functionality is multi-purpose in that it creates the network,
   when it doesn't exist.
2) If the DinD network exists, the parsing of the subnet fails,
   when IPv6 is involved (as there are multiple subnets).
3) Because of parsing issues, the gateway address calculation is
   wrong in IPv6 mode (uses IPv4 address).
4) Unrelated to PR 166, the NAT64 and DNS64 containers are
   deleted on "down" and restored on "up", dind::run will have
   already created these containers and duplicate containers
   occur.
5) IPv6 hasn't been tested for multi-cluster, and needs to be
   checked (and likely modified to dynamically create pod subnet
   based on the cluster instance), before being enabled.
6) The kube-router CI tests were failing, as well.
7) The test_src_master and test_src_release tests are failing in
   building binaries for test run.

For (1), we'll only determine the IP/prefix for the DinD network
once at startup. For IPv6, this is unchanged, it is either a
provided env variable, or default is used. For IPv4, it will
use the IP/prefix from the existing network, will use a provided
env variable value, or determine an available CIDR to use for the
DinD network and assign that to a variable for later use. This
separates network definition from creation.

For (2), the parsing is modified, to account for the possiblity
of an existing DinD network with an IPv6 subnet. This parsing
will only be done for IPv4 currently. In the future, IPv6 could
be parsed, and if no DinD network, an address can be calculated,
when not provided.

For (3), because we are using variables for the DinD network IP
and mask, and not looking up the values (and failing in the
parsing), the gateway assignment is trival.

For (4), during testing it was seen that NAT64 and DNS64 have the
same label as node containers, causing them to be deleted. The
solution is to use a different label, and only delete them during
a "clean" operation. The original goal was to reduce startup time,
by not creating these containers all the time. One a related note,
they likely can be shared by multiple clusters, so we don't want
to delete them during down.

For (5), a test was added at statup for IPv6 and mult-cluster, and
the script exited.

For (6), needed rebase with master to pick up fix to set context
for kube-router use. Update README.md to indicate how to use
kube-router.

For (7), commented out test_src_master and test_src_release test
cases in Circle CI, but later rebased to pick up PR #193, and un-
commented the tests.  Added quotes in one call, so that the
various binary names are not munged into one binary name for source
build.

Some future changes to be considered...

The network name is computed on every call that needs to reference
the kubeadm-dind-net. This too, could be done one and stored in a
variable.

The hash that is used for the label that determines the cluster,
could be a simple ID number. Unless we forsee a huge number of
clusters or some recyclying of cluster instances, a hash isn't
needed, IMHO. With a simple ID, it can be used in other places.

For example, the ID could be used in the network address for the
DinD network. For IPv6, one of the 16 bit parts of the network
part could be used. The same is true for IPv4, instead of trying
to find a free network. The subnet could be 172.<ID>.0.0/16. To
prevent a conflict with existing networks, an environment variable
could be used to specify the base prefix or minimum cluster ID. A
default value could match what is done today.

With IPv4, a calulation is used to determine IP addresses in the
DinD network. This could be reverted to the previous mechanism
that uses the last part of the dotted IP address for the interface
part. This would be fine as long as we have less than 255 IPs
needed per cluster (I can't imagine exceeding 10 IPs, frankly).

When IPv6 tests are added to CI, beside testing startup and operation,
the up/down/up cycle should be checked, along with persistence of
NAT64/DNS64 containers. To ensure that there is no regression for
changes made in (4). See issue #182 for IPv6 based CI test suggestions.