Refactor IPv6 tests
The main tests now are around name resolution and checking if the resolved IPv6
address is the public IP or the one we got from DNS64.

External ping tests are currently only running when DNS654 is in play and we
route traffic through NAT64 -- otherwise we will skip that tests.