Enable ability to use IPVS mode for kube-proxy
To be able to run kube-proxy in IPVS mode, the ipset app needs to be
present in the container. This commit adds this, and ipvsadm, and makes
them available on the nodes as well, so that IPVS status can be observed.

To run kube-proxy in IPVS mode, one needs to add the following to the
kubeadm.conf.tmpl file and rebuild the DinD image (and use it locally):

kubeProxy:
  config:
    mode: "ipvs"