IPv6: Multi-cluster support and dual-stack infrastructure
The primary goal of this commit is to put in place infrastructure to allow
kubeadm-dind-cluster to create clusters with dual-stack networks. Because
of the needed changes, it makes sense to make sure that dual-stack and IPv6
supports multi-cluster.

A secondary goal is to attempt to maintain backward compatibility with the
current mechanisms to customize and operate the cluster.

To prepare for dual-stack, the following changes were made (there will be
follow up commits to complete support for dual-stack, some in conjuction
with Kubernetes changes that will be made):

- Introduced 'dual-stack' IP MODE and adjusted existing IPv4 only and IPv6
  only logic to continue to operate correctly, when testing $IP_MODE.

- Introduced new MGMT_CIDRS env variable to allow a comma separated list of
  CIDRs to be provided for the mangement (dind) network. For IPv4/IPv6, a
  single CIDR will be specifed.

- For backwards compatibility, DIND_SUBNET and DIND_SUBNET_SIZE, if specified
  will be merged into MGMT_CIDRS.

- For multi-cluster, a new CLUSTER_ID will be used to allow the user to
  adapt the mgmt subnet used for each cluster. Removed logic that tries to
  determine the available subnet (see note 1).

- For compatibility, if (only) DIND_LABEL is provided, a random cluster ID
  will be assigned (see note 2).

- For additional clusters, in a multi-cluster setup, resource names will
  use the suffix "-cluster-#", instead of "-<label>-<sha#>".

- The localhost address, used for API, will be based on the family of the
  service network, which will be single family, when running in dual-stack
  mode. It is unchanged for IPv4/IPv6 only, which only have one family.

- During clean, it was found that the kubectl config for the context and
  cluster (in multi-cluster) are not removed. This commit removes those
  settings.

- For IPv6 only multi-cluster, separate DNS64 and NAT64 containers are
  created, especially because the docker bridge for the mgmt network
  needs to have a IPv4 network for the NAT64 IPv4/IPv6 mapping pool.
  This IPv4 network uses the cluster ID to ensure uniqueness. These
  containers are cleaned, based on the cluster context, which is encoded
  into thier label.

- Added multi-cluster CI test for IPv6 only (for 1.11 release). Adjusted
  portforward.sh to handle IPv6 mode. Using local address '127.0.0.1'
  and remote address '[::1]'.

- Modified existing multi-cluster CI test to check normal, label only
  (changed), and cluster ID (new) based clusters.

- Docker run commands are modified to define any IPs needed, based on
  the IP mode (two for dual-stack).

- Added cluster info printing, at the end of "up", so that user knows
  the network and cluster settings.

- Updated README.md to indicate the changes made.

- (FUTURE) If desired, could ensure that there is enough space in the
  mgmt CIDRs for the cluster ID to be injected.

- (FUTURE) For IPv6, could support the cluster ID only consuming 8 bits,
  instead of 16, and thus place it in either the upper or lower byte of
  the hextet, based on the CIDR size.

- (FUTURE) Could consider whether want to use /16 for mgmt CIDR in IPv4,
  allowing 32K nodes vs 254.

- (TODO) For multicluster IPv6 only, iptables rules are created on the
  host to isolate the clusters. These should be ip6tables rules. As a
  result, nodes from one cluster can currently ping nodes in another
  cluster (unless rules are manually added).

- (TODO) Complete the suport for dual-stack, including pod network, CNI,
  and other changes.

Note 1:
This came out of the need to support multiple clusters for IPv6, where
there are vast addresses and calculating an available subnet is difficult.
Picking a number to use to make the subnet unique, coupled with the base
selection via MGMT_CIDRS, provides a simple method to ensure uniqueness.

For IPv4, the management network will be /24, instead of /16, and the 3rd
octet of the subnet will be reserved for the cluster. For example, if the
MGMT_CIDRS is 10.192.0.0/24, and the CLUSTER_ID is "100", the management
CIDR will be modified to 10.192.100.0/24 for the cluster.

For IPv6, the CLUSTER_ID will be placed in the lowest hextet of the address
provided in MGMT_CIDRS (before the ::). For example, if the MGMT_CIDRS has
fd00:10:20::/64, and the CLUSTER_ID is "30", the management CIDR will be
modified to fd00:10:20:30::/64 for the cluster.

Values will be between 1..254 for IPv4 compatibility, with zero reserved for
single cluster or the first cluster in multi-cluster.

Each cluster would use a different cluster ID. The cluster ID will be used
in the names of resources, as well, to provide uniqueness.

Note 2:
Although kept for compatibility, this gives the opportunity to have a custom
cluster label, if desired. When only CLUSTER_ID is used, the cluster name will
be "cluster-#", where # is the non-zero cluster ID.

A cluster ID can be specified as well as the label, to control the mgmt
network address used.

Partially addresses issue: #209
Provides a solution for issue: #203