Version specific kubeadm.conf files
Changed this to use version specific config files and to use them
based on the kubeadm version, rather than having one template file
and trying to morph it for each version. Found that 1.12 is
considerably different config.

The 1.10 template is used for 1.8-1.10, and the 1.13 template is
used for 1.13 and newer versions.

Note: There are new sections and fields for the 1.12 an 1.13 kubeadm.conf
files. For 1.13, the conntack settings were commented out, as I suspect
they are host specific. Other, like the healthz ports, were left there,
even though they are showing IPv4 addresses. This template was generated
from a IPv6 setup, so, at this time, IPv4 is still used for some
things. These may need to be parameterize or removed later, as IPv6
changes are made.

Fixes Issue: #236