Add option to allow using pre-built Kubernetes binaries
Adding a "DIND_K8S_BIN_DIR" environment variable to allow the dind-cluster.sh
script to use pre-built binaries for hyperkube and kubeadm on the host.

For example, if the 'hyperkube' and 'kubeadm' binaries are in the directory
/home/myname/k8s_bin_dir on the host, then a user would set the environment
variable:

    export DIND_K8S_BIN_DIR=/home/myname/k8s_bin_dir

and then run 'dind-cluster.sh up' in the normal fashion.

With the proposed change, dind-cluster.sh would create a bind mount from
that directory on the host to the /k8s directory on each node container.

Note that the setting of DIND_K8S_BIN_DIR to a non-null string will cause
the BUILD_HYPERKUBE and BUILD_KUBEADM environment variables to be
ignored.